requirements.txt
# Sifra Dependencies

ollama>=0.1.0
chromadb>=0.4.0
pyttsx3>=2.90
openai-whisper>=20231117
beautifulsoup4>=4.12.0
requests>=2.31.0
rich>=13.7.0
prompt_toolkit>=3.0.43
PyYAML>=6.0.1
# pyaudio>=0.2.14 # Uncomment for voice support (requires portaudio dev headers)
